Add files via upload
This C++ program prompts the user to enter a number and then checks whether the number is positive or negative. It uses an      "if-else" statement to compare the entered number with zero. If the input statement/condition is true then "if" statement is executed and if the condition is false then "else" statement is executed.