Add files via upload
This C++ program demonstrates how to work with strings. It defines a string variable "str" and assigns it the value "Hello World". The program then uses the "cout" statement to output the contents of the string to the screen. Finally, it ends by returning "0", indicating that the program has executed successfully. This basic example illustrates how to handle and print string data in C++